Make Child app from Parent app with modules

Make Child app from Parent app with modules

Usage

make_child_app(
  parent_app_dir,
  child_app_dir,
  include_renv = FALSE,
  copy_jobs_dir = FALSE,
  include_meta_yaml = TRUE,
  job_file_type = "Rmd",
  framework = "none",
  overwrite = FALSE
)

Arguments

parent_app_dir

character Path to existing parent app, including the parent app directory

child_app_dir

character Path to new child app, including the child app directory

include_renv

logical Include renv structure in child app or now. Default is FALSE

copy_jobs_dir

logical Copy the jobs directory from the parent app (TRUE) or use jobs template in matte/inst. Default is FALSE

include_meta_yaml

logical Include the template metadata yaml file in the child app. Default is TRUE

job_file_type

character Type of job template to include ("Rmd" or "qmd"). Default is Rmd

framework

character App framework used for child. Options are "golem", "rhino" or "none". Default is "none"

overwrite

logical Allow function to overwrite existing child directory. Default is FALSE

Examples

if (FALSE) { # \dontrun{
make_child_app(parent_app_dir = "inst/parentApp",
               child_app_dir = "~/childTest")

make_child_app(parent_app_dir = "inst/parentApp",
               child_app_dir = "~/childTest2",
               copy_jobs_dir = FALSE,
               include_renv = TRUE)

make_child_app(parent_app_dir = "inst/parentApp",
               child_app_dir = "~/childTest3",
               include_renv = TRUE,
               framework = "golem")
} # }